requirements.txt
python-dotenv==1.0.0
aiogram==2.25.1
Flask==2.3.3
telethon==1.29.2
requests==2.31.0 